+++ R/footnote.R
@@ -107,12 +107,10 @@ footnote <- function(x, i = NULL, j = NULL, value, ref_symbols = NULL, part = "b
   if (any(symbols_str %in% "")) {
     long_msg <- c(
       "Usage of empty symbol '' with footnote should not happen, ",
-      "use `add_footer_lines()` instead, it does not require any symbol. ",
-      "This usage will be forbidden in the next release. Please, wait for 10 seconds!"
+      "use `add_footer_lines()` instead, it does not require any symbol. "
     )
     long_msg <- paste0(long_msg, collapse = "\n")
-    message(long_msg)
-    Sys.sleep(10)
+    stop(long_msg)
   }
 
   sep_str <- rep(sep, length(value))